moved ChargerStationTableRow to its own file
